backend/requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-docx==1.1.0
PyPDF2==3.0.1
beautifulsoup4==4.12.2
requests==2.31.0
httpx==0.25.0
python-dotenv==1.0.0
python-multipart==0.0.6
reportlab==4.2.5